backtesting-service/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
httpx==0.19.0
pydantic==1.8.2
python-dotenv==0.19.0
numpy==1.21.2
pandas==1.3.3
matplotlib==3.4.3
scikit-learn==0.24.2
pytest==7.3.1
pytest-asyncio==0.21.0
pytest-mock==3.10.0
redis==5.0.1
aioredis==2.0.1